add complaint

+++ src/Components/CartView/CartItemView/CartItemView.js
@@ -1,14 +1,15 @@
 import {View,usePressGesture} from "corella";
 import css from "./CartItemView.module.css";
+import ConfirmationView from "./ConfirmationView/ConfirmationView";
 import {minus0,plus0,trash0} from "assets";
 
 
 export default function CartItemView(props){
     const {parent,cartitem,onChangeQuantity,onRemove}=props;
-    const cartitemview=View({parent,className:css.cartitemview});
+    let cartitemview=View({parent,className:css.cartitemview});
 
     cartitemview.innateHTML=`
-        <div class="${css.col0}">
+        <div class="${css.col0}" ref="col0">
             <text class="${css.name}">${cartitem.product.name}</text>
             <text class="${css.price}">
                 ${cartitem.granularity} - ${language[cartitem.unit==="t"?"ton":"kilogram"]}
@@ -49,8 +50,30 @@ export default function CartItemView(props){
     });
 
     cartitemview.removebtn.onclick=()=>{
-        cartitemview.remove();
-        onRemove&&onRemove(cartitem);
+        ConfirmationView({
+            parent:cartitemview,
+            onConfirm:()=>{
+                cartitemview.remove();
+                onRemove&&onRemove(cartitem);
+            },
+        });
+    }
+
+    cartitemview.col0.onclick=()=>{
+        const {product}=cartitem;
+        WebView.show({
+            id:"bottomsheet",
+            message:{
+                title:product.name,product,
+                contentId:"productorder",
+            },
+            onClose:({store})=>{
+                const {cart}=store;
+                const item=cart.items.find(item=>item.product.id===product.id);
+                Object.assign(cartitem,item);
+                cartitemview=cartitemview.substitute(CartItemView(props));
+            },
+        });
     }
